refactor(footer) replace footerLinks into utils
